jarvis/pyproject.toml
[project]
name = "jarvis"
version = "0.1.0"
description = "JARVIS - Modular AI Personal Assistant"
authors = [{ name = "Steve Penn" }]
requires-python = ">=3.9"
dependencies = [
    "ollama>=0.3.0",
    "faster-whisper>=1.0.0",
    "elevenlabs>=1.0.0",
    "pyaudio>=0.2.14",
    "pvporcupine>=3.0.0",
    "pyyaml>=6.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "rich>=13.0.0",
    "flet>=0.21.0",
    "opencv-python>=4.9.0",
    "mss>=9.0.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0"]

[project.scripts]
jarvis = "jarvis.cli:app"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
